refactor/ImageMapper receives props as separeted file to the parent floors.js
