+++ src/components/LocationComponents/Floors.js
@@ -1,20 +1,10 @@
-import React, { useState } from 'react'
+import React from 'react'
 import Navigation from './Navigation'
 import data from './FloorsMapData'
-import ImageMapper from 'reactjs-img-mapper'
+import ImageMap from './ImageMap'
 import "./Mapper.css"
 
 export const Floors = () => {
-    const [hoveredArea, setHoveredArea] = useState(null)
-
-    const handleMouseEnter = (area, evt) => {
-        const { clientX, clientY } = evt
-        setHoveredArea({ ...area, clientX, clientY })
-    }
-
-    const handleMouseLeave = () => {
-        setHoveredArea(null)
-    }
 
     return (
         <section className="relative w-full min-h-screen flex items-center justify-center bg-gradient-to-l from-[#034687] to-[#00c379]">
@@ -23,8 +13,6 @@ export const Floors = () => {
                 <div className="w-fit h-fit md:py-8 gap-4 lg:gap-14 grid xl:grid-cols-3 lg:grid-cols-2 grid-cols-1 grid-flow-row justify-self-auto place-items-center mx-[20px] my-[120px] auto-cols-auto">
                     {data.map((item, index) => {
                         const imgCoords = data.find((coord) => coord.name === item.name);
-                        const currentImage = hoveredArea && hoveredArea.item === item.name
-
                         return (
                             <div key={index} className="relative m-auto">
                                 <span className="flex items-center justify-center gap-2">
@@ -33,32 +21,7 @@ export const Floors = () => {
                                     </svg>
                                     <h3 className="text-xl md:text-2xl text-black leading-7 font-medium font-serif mb-2">{item.title}</h3>
                                 </span>
-                                <ImageMapper
-                                    src={item.imgUrl}
-                                    map={imgCoords}
-                                    imgWidth={500}
-                                    width={500}
-                                    height={400}
-                                    onMouseEnter={handleMouseEnter}
-                                    onMouseLeave={handleMouseLeave}
-                                />
-                                {hoveredArea && setHoveredArea.name !== item.name && (
-                                    <div
-                                        style={{
-                                            zIndex: "99",
-                                            position: 'absolute',
-                                            top: hoveredArea.clientY - 100,
-                                            left: hoveredArea.clientX - 50,
-                                            transform: 'translate(-50%, -50%)',
-                                            background: hoveredArea.preFillColor || 'rgba(0, 0, 0, 0.7)',
-                                            color: hoveredArea.fillColor || 'white',
-                                            padding: '5px',
-                                            fontWeight: 'bold',
-                                        }}
-                                    >
-                                        {hoveredArea.name}
-                                    </div>
-                                )}
+                                <ImageMap item={item} map={imgCoords} />
                             </div>
                         );
                     })}
